myproject/requirements.txt
boto3
clickhouse-driver
djangorestframework
clickhouse-connect
dotenv
pandas